Implement balanced uniform positioning

+++ riskzones.py
@@ -17,12 +17,29 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
+from enum import Enum
 import osmpois
 import json
 import sys
 import os
 import random
 import numpy
+
+'''
+Exception classes.
+'''
+class OutOfBounds(Exception):
+    pass
+
+class SkipZone(Exception):
+    pass
+
+'''
+Positioning modes.
+'''
+class UniformPositioningMode(Enum):
+    UNBALANCED = 1
+    BALANCED = 2
 
 '''
 This class contains the methods for calculating zone risks from PoIs
@@ -78,7 +95,8 @@ class RiskZonesGrid:
                     'lon': (i / self.grid_x * self.width) + self.left + self.zone_center['x'],
                     'risk': 1.0,
                     'RL': 0,
-                    'inside': True
+                    'inside': True,
+                    'has_edu': False
                 }
 
                 self.zones.append(zone)
@@ -301,7 +319,7 @@ class RiskZonesGrid:
     '''
     Uniformly select zones for EDUs positioning.
     '''
-    def set_edus_positions_uniform(self):
+    def set_edus_positions_uniform(self, mode: UniformPositioningMode):
         edus = self.__get_number_of_edus_by_RL()
         self.edus = {}
         At = {}
@@ -312,41 +330,119 @@ class RiskZonesGrid:
         step_y = {}
         zone_in_y = {}
         start = {}
+        total = len(self.zones)
+        print('\nPositioning EDUs...', end='\r')
+
+        # Reset zones EDUs flag
+        for zone in self.zones:
+            zone['has_edu'] = False
 
         for i in range(1, self.M + 1):
-            At[i] = self.__get_number_of_zones_by_RL()[i]  # Area of the whole RL
-            Ax[i] = numpy.ceil(At[i] / edus[i])            # Coverage area of an EDU
-            radius[i] = numpy.ceil(numpy.sqrt(Ax[i]) / 2)  # Radius of an EDU
-            step[i] = 2 * radius[i]                        # Step distance on x and y directions
-            start[i] = radius[i]                           # Start coordinate
-            self.edus[i] = []                              # Final list of EDUs in zone i
-            step_x[i] = step_y[i] = 0                      # The steps are accounted individually for each RL
-            zone_in_y[i] = False                           # To check if there was any zone for a RL in any y
+            At[i] = self.__get_number_of_zones_by_RL()[i]        # Area of the whole RL
+            Ax[i] = numpy.ceil(At[i] / edus[i])                  # Coverage area of an EDU
+            radius[i] = numpy.ceil(numpy.sqrt(Ax[i]) / numpy.pi) # Radius of an EDU
+            step[i] = 2 * radius[i]                              # Step distance on x and y directions
+            start[i] = radius[i]                                 # Start coordinate
+            self.edus[i] = []                                    # Final list of EDUs in zone i
+            step_x[i] = step_y[i] = 0                            # The steps are accounted individually for each RL
+            zone_in_y[i] = False                                 # To check if there was any zone for a RL in any y
+        smallest_radius = int(radius[self.M])                    # Radius of the highest level
+        highest_radius = int(radius[1])                          # Radius of the lowest level
         
         self.zones.sort(key=lambda zone : zone['id'])
-        for y in range(self.grid_y):
-            # First, reset step for every RL in x direction and check if there was any zone in y
-            for i in range(1, self.M + 1):
-                step_x[i] = 0
-                if zone_in_y[i]:
-                    step_y[i] += 1
-                    zone_in_y[i] = False
 
-            # For each zone in this coordinate, check if it is inside AoI and check if it is time to
-            # put an EDU in it
-            for x in range(self.grid_x):
-                id = self.grid_x * y + x
-                zone = self.zones[id]
-                if not zone['inside']: continue
-
+        #
+        # Unbalanced positioning mode.
+        #
+        if mode == UniformPositioningMode.UNBALANCED:
+            for y in range(self.grid_y):
+                # First, reset step for every RL in x direction and check if there was any zone in y
                 for i in range(1, self.M + 1):
-                    if zone['RL'] != i: continue
-                    zone_in_y[i] = True # If there was any zone for this RL in this y, we can increment step_y later
+                    step_x[i] = 0
+                    if zone_in_y[i]:
+                        step_y[i] += 1
+                        zone_in_y[i] = False
 
-                    if step_x[i] % step[i] == 0 and step_y[i] % step[i] == 0:
-                        self.edus[i].append(zone)
+                # For each zone in this coordinate, check if it is inside AoI and check if it is time to
+                # put an EDU in it
+                for x in range(self.grid_x):
+                    id = self.grid_x * y + x
+                    zone = self.zones[id]
+                    if not zone['inside']: continue
+
+                    for i in range(1, self.M + 1):
+                        if zone['RL'] != i: continue
+                        zone_in_y[i] = True # If there was any zone for this RL in this y, we can increment step_y later
+
+                        if step_x[i] % step[i] == 0 and step_y[i] % step[i] == 0:
+                            self.edus[i].append(zone)
+                            
+                        step_x[i] += 1
+
+                        prog = (id / total) * 100
+                        print(f'Positioning EDUs... {prog:.2f}%', end='\r')
+        
+        #
+        # Balanced positioning mode.
+        #
+        elif mode == UniformPositioningMode.BALANCED:
+            y = 0
+            while y < self.grid_y:
+                x = 0
+                try:
+                    while x < self.grid_x:
+                        while True:
+                            id = self.grid_x * y + x
+                            zone = self.zones[id]
+                            if zone['inside']:
+                                break
+                            elif x >= self.grid_x:
+                                raise OutOfBounds
+                            else:
+                                x += 1
+
+                        try:
+                            # Don't even try if we are still within the range of another EDU
+                            nearby_zones = self.__get_zones_in_area(id, 2 * highest_radius + 1)
+                            for nearby_zone in nearby_zones:
+                                if not nearby_zone['has_edu']: continue
+                                calc_radius = 2 * radius[zone['RL']] * self.zone_size
+                                if self.__calculate_distance(zone, nearby_zone) <= (calc_radius):
+                                    raise SkipZone
+
+                            self.edus[zone['RL']].append(zone)
+                            zone['has_edu'] = True
+                            x += smallest_radius
                         
-                    step_x[i] += 1
+                        except SkipZone:
+                            x += 1
+                    
+                        prog = (id / total) * 100
+                        print(f'Positioning EDUs... {prog:.2f}%', end='\r')
+
+                except IndexError:
+                    pass
+                except OutOfBounds:
+                    pass
+                
+                y += smallest_radius
+
+        print(f'Positioning EDUs... 100.00%')
+
+    '''
+    Get all zones within a squared area.
+    '''
+    def __get_zones_in_area(self, center_id: int, radius: int) -> list:
+        start_x = int(center_id % self.grid_x) - radius
+        start_y = int(center_id / self.grid_x) - radius
+        zones = []
+
+        for i in range(start_y, radius * 2 + start_y + 1):
+            for j in range(start_x, radius * 2 + start_x + 1):
+                zones.append(self.zones[i * self.grid_x + j])
+        
+        zones.sort(key=lambda zone : zone['id'])
+        return zones
 
 '''
 Main program.
@@ -396,7 +492,7 @@ if __name__ == '__main__':
         grid.calculate_risk_from_pois(pois)
 
     # Write cache file
-    if conf['cache_zones'] == True:
+    if conf['cache_zones'] == True and not os.path.isfile(cache_filename):
         fp = open(cache_filename, 'w')
         json.dump(grid.zones, fp)
         fp.close()
@@ -417,8 +513,7 @@ if __name__ == '__main__':
     fp.close()
     
     # Write a CSV file with EDUs positioning
-    #grid.set_edus_positions_random()
-    grid.set_edus_positions_uniform()
+    grid.set_edus_positions_uniform(UniformPositioningMode.UNBALANCED)
     row = 0
     data = ''
     data += 'system:index,.geo\n'
